backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
python-multipart==0.0.6
psycopg2-binary==2.9.9
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

